requirements.txt
openai>=1.12.0
openai-whisper>=20231117
pydantic>=2.0.0
python-dotenv>=1.0.0
pyyaml>=6.0.0

soundfile>=0.12.1
scipy>=1.11.0
numpy>=1.24.0

sentence-transformers>=2.2.0
scikit-learn>=1.3.0

aiofiles>=23.2.0
typer>=0.9.0

pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
ruff>=0.1.0
mypy>=1.7.0
